+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- A new kwarg `unique_workdir` that will create unique subfolders of the type `<DISPATCH ID>/node_<NODE ID>` within `remote_workdir` if set to `True`
+
+### Fixed
+
+- Fixed a bug where `cleanup = False` would be ignored.
+- Moved Oathtool import check to the `init` for quicker feedback
+- Increased overall test coverage
+
 ## [0.12.1] - 2023-05-05
 
 ### Fixed